% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_from_db.R
\name{database.to.graph}
\alias{database.to.graph}
\title{Creates an igraph object from a thread in the database}
\usage{
database.to.graph(idthread, con, dataset)
}
\arguments{
\item{idthread}{thread id}

\item{con}{RSQlite connection to a database}

\item{dataset}{name of the forum}
}
\value{
a tree graph of the thread
}
\description{
Creates an igraph object from a thread in the database
}

