% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimators.R
\name{estimation_Lumbreras2016}
\alias{estimation_Lumbreras2016}
\title{Expectation-Maximization algorithm to find clusters and 
parameters of each cluster}
\usage{
estimation_Lumbreras2016(df.trees, params, niters = 10)
}
\arguments{
\item{df.trees}{dataframe of posts}

\item{params}{list of initial parameters}
}
\value{
list of final parameters
}
\description{
Expectation-Maximization algorithm to find clusters and 
parameters of each cluster
}

