% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{generate_trees}
\alias{generate_trees}
\title{Generate n trees of sizes s following according to a model with given parameters}
\usage{
generate_trees(model, params, sizes)
}
\arguments{
\item{model}{likelihood model that drives parent choice for every post}

\item{params}{a vector of model parameters}

\item{n}{total number of threads}

\item{s}{size of each thread}
}
\value{
a list of trees
}
\description{
Generate n trees of sizes s following according to a model with given parameters
}
\details{
The trees are generated in parallel
}

