% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihood.R
\name{likelihood.Gomez2013.deprecated}
\alias{likelihood.Gomez2013.deprecated}
\title{Deprecated b/c it uses igraph keep it just for testing}
\usage{
likelihood.Gomez2013.deprecated(trees, alpha = 1, beta = 1, tau = 0.75)
}
\arguments{
\item{trees}{list of trees}

\item{alpha}{popularity parameter}

\item{beta}{root bias}

\item{tau}{recency parameter}
}
\value{
loglikelihood of the dataset
}
\description{
Deprecated b/c it uses igraph keep it just for testing
}

