% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihood.R
\name{likelihood_Gomez2013}
\alias{likelihood_Gomez2013}
\title{Total likelihood of a dataframe according to Gomez2013}
\usage{
likelihood_Gomez2013(df.trees, params)
}
\arguments{
\item{df.trees}{data.frame with one post per row and features in columns}

\item{model}{parameters}
}
\value{
loglikelihood of the dataset
}
\description{
Total likelihood of a dataframe according to Gomez2013
}

