% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihood.R
\name{likelihood_Lumbreras2016}
\alias{likelihood_Lumbreras2016}
\title{Total likelihood of a dataframe according to Lumbreras2016}
\usage{
likelihood_Lumbreras2016(df.trees, params, responsabilities, pis)
}
\arguments{
\item{df.trees}{data.frame with one post per row and features in columns}

\item{responsabilities}{responsabilities of users w.r.t clusters}

\item{model}{parameters}
}
\value{
loglikelihood of the dataset
}
\description{
Total likelihood of a dataframe according to Lumbreras2016
}

