% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihood.R
\name{likelihood_post}
\alias{likelihood_post}
\title{Likelihood computation using the dataframe}
\usage{
likelihood_post(row, alpha, beta, tau)
}
\arguments{
\item{row}{vector representing a post}

\item{alpha}{popularity parameter}

\item{beta}{root bias}

\item{tau}{recency parameter}
}
\value{
loglikelihood of the post
}
\description{
Likelihood computation using the dataframe
}

