% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{load_trees}
\alias{load_trees}
\title{Build a set of tree objects from the information in the database}
\usage{
load_trees(forum = "reddit", subforum = "gameofthrones")
}
\arguments{
\item{forum}{the forum (reddit, slashdot,...)}

\item{subforum}{the subforum (topic)}
}
\value{
a list of trees
}
\description{
Build a set of tree objects from the information in the database
}
\details{
This function builds a set of tree objects from the information stored in the databased.
For instance, if forum='reddit' and subforum='gameofthrones' then it will extract all posts in the 
reddit database that belong to gameofthrones and will build a list of trees where every tree represents 
a discussion thread in gameofthrones.
}

