% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_from_db.R
\name{plot.threadgraph}
\alias{plot.threadgraph}
\title{Plots the user and the tree posts graph of a thread}
\usage{
\method{plot}{threadgraph}(threadgraph, title = "")
}
\arguments{
\item{threadgraph}{a list with a $gu (users) graph and $gp (posts) graph}

\item{title}{plot title}
}
\description{
Plots the user and the tree posts graph of a thread
}

