% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihood.R
\name{Qopt}
\alias{Qopt}
\title{The part of the lower bound that we optimize in the M-step}
\usage{
Qopt(params, df.trees, responsabilities, pis, k)
}
\description{
The part of the lower bound that we optimize in the M-step
}

